Move suggestionToken to a prop in RegisterDomainStep.
